backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.42.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.3.0
python-multipart==0.0.9
regex==2024.5.15
numpy==1.26.4
scikit-learn==1.5.0
